+++ model/src/main/kotlin/io/github/openminigameserver/replay/model/recordable/RecordablePosition.kt
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.annotation.JsonInclude
 
 @JsonInclude(JsonInclude.Include.NON_DEFAULT)
 data class RecordablePosition(
-    val x: Float = 0f,
-    val y: Float = 0f,
-    val z: Float = 0f,
+    val x: Double = 0.0,
+    val y: Double = 0.0,
+    val z: Double = 0.0,
     val yaw: Float = 0f,
     val pitch: Float = 0f
 )
